har gjort en readme fil samt gjort en liten ändring i post och valideringe
