+++ serverSrc/data/validationCart.ts
@@ -1,25 +1,25 @@
-// import { z } from "zod";
+import { z } from "zod";
 
-// // cartItemSchema -> validerar VARJE object i kundvagnen
-// // productId -> måste vara en sträng, som inte heller får vara tom
-// // amount -> måste vara ett POSITIVT HELTAL
-// export const cartItemSchema = z.object({
-//   productId: z.string().nonempty("productId krävs"),
-//   amount: z.number().int().positive("amount måste vara ett positivt heltal"),
-// });
+// cartItemSchema -> validerar VARJE object i kundvagnen
+// productId -> måste vara en sträng, som inte heller får vara tom
+// amount -> måste vara ett POSITIVT HELTAL
+export const cartItemSchema = z.object({
+  productId: z.string().nonempty("productId krävs"),
+  amount: z.number().int().positive("amount måste vara ett positivt heltal"),
+});
 
-// // Här validerar vi hela kundvagnen som en array av cartItemSchema
+// Här validerar vi hela kundvagnen som en array av cartItemSchema
 
-// export const cartSchema = z.object({
-//   id: z.string().nonempty("id krävs"),
-//   userId: z.string().nonempty("userId krävs"),
-//   productId: z.string().nonempty("productId krävs"),
-//   amount: z.number().int().positive("amount måste vara ett positivt heltal"),
-// });
+export const cartSchema = z.object({
+  id: z.string().nonempty("id krävs"),
+  userId: z.string().nonempty("userId krävs"),
+  productId: z.string().nonempty("productId krävs"),
+  amount: z.number().int().positive("amount måste vara ett positivt heltal"),
+});
 
-// export const cartsSchema = z.array(cartSchema);
+export const cartsSchema = z.array(cartSchema);
 
-// // Vi genererar TS-typer automatiskt från schemat
+// Vi genererar TS-typer automatiskt från schemat
 
-// export type Cart = z.infer<typeof cartItemSchema>;
-// export type Carts = z.infer<typeof cartsSchema>;
+export type Cart = z.infer<typeof cartItemSchema>;
+export type Carts = z.infer<typeof cartsSchema>;
